# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Management Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/auth": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Create a new user",
                "description": "Create a new user with the provided data",
                "operationId": "8fdc39ba963c089d882406b6374fc7f7",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "mobile"
                                ],
                                "properties": {
                                    "first_name": {
                                        "description": "User's first name",
                                        "type": "string",
                                        "maxLength": 30,
                                        "example": "milad",
                                        "nullable": true
                                    },
                                    "last_name": {
                                        "description": "User's last name",
                                        "type": "string",
                                        "maxLength": 30,
                                        "example": "rezai",
                                        "nullable": true
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string",
                                        "format": "password",
                                        "maxLength": 12,
                                        "minLength": 8,
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "description": "User's password",
                                        "type": "string",
                                        "format": "password",
                                        "maxLength": 12,
                                        "minLength": 8,
                                        "example": "secret123"
                                    },
                                    "mobile": {
                                        "description": "User's mobile number",
                                        "type": "string",
                                        "example": "09167516826"
                                    },
                                    "profile_photo_path": {
                                        "description": "User's frofile photo",
                                        "type": "file:png,jpg,jpeg",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "milad"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "rezai"
                                                },
                                                "mobile": {
                                                    "type": "string",
                                                    "example": "09167516826"
                                                },
                                                "profile_photo_path": {
                                                    "type": "string",
                                                    "example": "images/users/2025/12/01/1764590305.png"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "2|vlLCKWNJw1tDEsGYhUY6YqS3egSdfvfKr1EXG7NM0abe24e2"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "login a  user",
                "description": "login a  user with the provided data",
                "operationId": "3b03eb66e9fc14dbed91d82e3a240a06",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "mobile"
                                ],
                                "properties": {
                                    "password": {
                                        "description": "User's password",
                                        "type": "string",
                                        "format": "password",
                                        "maxLength": 12,
                                        "minLength": 8,
                                        "example": "secret123"
                                    },
                                    "mobile": {
                                        "description": "User's mobile number",
                                        "type": "string",
                                        "example": "09167516826"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "milad"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "rezai"
                                                },
                                                "mobile": {
                                                    "type": "string",
                                                    "example": "09167516826"
                                                },
                                                "profile_photo_path": {
                                                    "type": "string",
                                                    "example": "images/users/2025/12/01/1764590305.png"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "2|vlLCKWNJw1tDEsGYhUY6YqS3egSdfvfKr1EXG7NM0abe24e2"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "type": "",
                                                    "example": ""
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "data invalid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "logout a  user",
                "description": "logout a  user with the token",
                "operationId": "ee643cac55a982fc398983af84bf28aa",
                "responses": {
                    "204": {
                        "description": "User logout successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}